[optimize]: cleaning unwanted source and header files
