+++ client/integrationClient.cpp
@@ -21,6 +21,45 @@ public:
 
 #endif //PLAYGROUND
 
+#ifdef FUNCTION_PTR
+
+void func1() {
+	cout << "inside func1" << endl;
+}
+
+int simple_sum(int a, int b) {
+	return a + b;
+}
+
+void another1() {
+	cout << "another1" << endl;
+}
+
+void fn_pointer() {
+
+	cout << "Direct call: " << endl;
+	func1();
+
+	cout << "Using fp: " << endl;
+	void(*pf) () = &func1;
+	(*pf)();
+
+	void(*p_another) () = pf;
+	pf = &another1;
+	cout << "After assignment of fp to p_another" << endl;
+	cout << "using p_another" << endl;
+	(*p_another)();
+	cout << "using pf " << endl;
+	(*pf)();
+	cout << endl;
+
+	cout << "another using fp2" << endl;
+	int(*pf2)(int, int) = &simple_sum;
+	cout << "sum " << (*pf2)(3, 2) << endl;
+}
+
+#endif //FUNCTION_PTR
+
 
 int main() {
 
@@ -90,14 +129,20 @@ int main() {
 	//Templates and Metaprogramming
 	Derived <int> derived;
 	derived.display();
-	cout << "cube from Derived::cube_int(): " << derived.cube_int(5) << endl;
+	//cout << "cube from Derived::cube_int(): " << derived.cube_int(5) << endl;
 
 	Templatized <Int_class, Float_class> templatized;
 	Int_class* intclass = new Int_class();
 	Float_class* floatclass = new Float_class();
 
-	cout << "is int class object size is bigger: "
-		<< templatized.is_bigger_object_size(*intclass, *floatclass) << endl;
+	//cout << "is int class object size is bigger: "
+	//	<< templatized.is_bigger_object_size(*intclass, *floatclass) << endl;
+
+	Derived <float> d_float; 
+	d_float.display(); 
+
+
+
 #endif //TEMPLATE
 
 #ifdef QUEUE
