--- inc/algorithm.h
@@ -1,8 +1,0 @@
-#pragma once
-#include 
-
-namespace kumar {
-
-
-	int count_if(vector <int> input) {}
-}
