--- inc/calc.h
@@ -1,8 +1,0 @@
-#pragma once 
-#ifndef CALC_H
-#define CALC_H
-
-//__declspec(dllexport) int PullInMyLibrary() { return 0; }
-double avg_salary(int exp);
-
-#endif
