--- inc/fp.hpp
@@ -1,9 +1,0 @@
-#pragma once
-#ifndef FP_HPP
-#define FP_HPP
-
-#include "my_project_app.h"
-
-void fn_pointer();
-
-#endif //FP_HPP
