+++ inc/my_project_app.h
@@ -49,7 +49,6 @@
 #include "btree.hpp"
 #include "dp.hpp"
 #include "Threads.hpp"
-#include "fp.hpp"
 #include "Templates.hpp"
 #include "Queue_Array.hpp"
 #include "Observer.hpp"
