--- source/calc.cpp
@@ -1,6 +1,0 @@
-#include "calc.h"
-
-double avg_salary (int exp) {
-	
-	return 2.5 * double (exp);
-}
